CPU Time optimization : replace all double AHRS data by float. Creation AHRS thread.

+++ main/MadgwickAHRS.c
@@ -30,12 +30,12 @@
 //---------------------------------------------------------------------------------------------------
 // AHRS algorithm update
 */
-void MadgwickAHRSupdate(double dt,double gx, double gy, double gz, double ax, double ay, double az, double mx, double my, double mz, double *pq0, double *pq1, double *pq2, double *pq3) {
-	double recipNorm;
-	double s0, s1, s2, s3;
-	double qDot1, qDot2, qDot3, qDot4;
-	double hx, hy;
-	double _2q0mx, _2q0my, _2q0mz, _2q1mx, _2bx, _2bz, _4bx, _4bz, _2q0, _2q1, _2q2, _2q3, _2q0q2, _2q2q3, q0q0, q0q1, q0q2, q0q3, q1q1, q1q2, q1q3, q2q2, q2q3, q3q3;
+void MadgwickAHRSupdate(float dt,float gx, float gy, float gz, float ax, float ay, float az, float mx, float my, float mz, float *pq0, float *pq1, float *pq2, float *pq3) {
+	float recipNorm;
+	float s0, s1, s2, s3;
+	float qDot1, qDot2, qDot3, qDot4;
+	float hx, hy;
+	float _2q0mx, _2q0my, _2q0mz, _2q1mx, _2bx, _2bz, _4bx, _4bz, _2q0, _2q1, _2q2, _2q3, _2q0q2, _2q2q3, q0q0, q0q1, q0q2, q0q3, q1q1, q1q2, q1q3, q2q2, q2q3, q3q3;
 
 	/* Use IMU algorithm if magnetometer measurement invalid (avoids NaN in magnetometer normalisation)*/
 	if((mx == 0.0f) && (my == 0.0f) && (mz == 0.0f)) {
@@ -129,11 +129,11 @@ void MadgwickAHRSupdate(double dt,double gx, double gy, double gz, double ax, do
 /*---------------------------------------------------------------------------------------------------
 // IMU algorithm update
 */
-void MadgwickAHRSupdateIMU(double dt,double gx, double gy, double gz, double ax, double ay, double az, double *pq0, double *pq1, double *pq2, double *pq3) {
-	double recipNorm;
-	double s0, s1, s2, s3;
-	double qDot1, qDot2, qDot3, qDot4;
-	double _2q0, _2q1, _2q2, _2q3, _4q0, _4q1, _4q2 ,_8q1, _8q2, q0q0, q1q1, q2q2, q3q3;
+void MadgwickAHRSupdateIMU(float dt,float gx, float gy, float gz, float ax, float ay, float az, float *pq0, float *pq1, float *pq2, float *pq3) {
+	float recipNorm;
+	float s0, s1, s2, s3;
+	float qDot1, qDot2, qDot3, qDot4;
+	float _2q0, _2q1, _2q2, _2q3, _4q0, _4q1, _4q2 ,_8q1, _8q2, q0q0, q1q1, q2q2, q3q3;
 
 //	ESP_LOGI(FNAME, "Début MAg:%f,%f,%f,%f,%f,%f ", gx,ax,*pq0,*pq1,*pq2,*pq3);
 	/* Rate of change of quaternion from gyroscope*/
@@ -202,12 +202,12 @@ void MadgwickAHRSupdateIMU(double dt,double gx, double gy, double gz, double ax,
 // Fast inverse square-root
 // See: http://en.wikipedia.org/wiki/Fast_inverse_square_root
 
-double invSqrt(double x) {
-	double halfx = 0.5f * x;
-	double y = x;
-	long i = *(double*)&y;
+float invSqrt(float x) {
+	float halfx = 0.5f * x;
+	float y = x;
+	long i = *(float*)&y;
 	i = 0x5f3759df - (i>>1);
-	y = *(double*)&i;
+	y = *(float*)&i;
 	y = y * (1.5f - (halfx * y * y));
 	return (y);
 }
